Fixes display bug of chart
The chart diagram was flickering. The root cause of this was, that if a
new chart was requested from the selection menu, it was placed on top of
the existing chart. Now if a new chart is requested, the already
existing chart is destroyed and a new one is created.